Move classes to local folder
