+++ classes/local/banner.php
@@ -22,7 +22,7 @@
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace local_softwarewarning;
+namespace local_softwarewarning\local;
 
 /**
  * Banner class
