+++ classes/task/download_browscap.php
@@ -22,7 +22,7 @@
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace local_softwarewarning\task;
+namespace local_softwarewarning\local\task;
 
 /**
  * Softwarewarning download browscap
