+++ lib.php
@@ -13,7 +13,7 @@ function local_softwarewarning_before_http_headers() {
     if ($PAGE->theme->name !== 'wwu2019') {
         return;
     }
-    $banner = \local_softwarewarning\cache_util::get_banner();
+    $banner = \local_softwarewarning\local\banner_manager::get_banner();
     if ($banner) {
         $PAGE->add_body_class('withbanner');
     }
@@ -24,7 +24,7 @@ function local_softwarewarning_before_standard_top_of_body_html(): string {
     if ($PAGE->theme->name !== 'wwu2019') {
         return '';
     }
-    $banner = \local_softwarewarning\cache_util::get_banner();
+    $banner = \local_softwarewarning\local\banner_manager::get_banner();
     if (!$banner) {
         return '';
     }
